Move RankingEvaluator to AUCEvaluator
